pyproject.toml
[tool.poetry]
name = "doc-processing"
version = "0.1.0"
description = ""
authors = ["Tamir Mandreev <tamirmandreev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.2"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
ipython = "^9.3.0"
pytest = "^8.4.0"
pytest-django = "^4.11.1"
celery = "^5.5.3"
redis = "^6.2.0"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.1"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.7.0"
black = "^25.1.0"
flake8 = "^7.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
